@@ -18,22 +18,22 @@
     "lint": "ng lint"
   },
   "dependencies": {
-    "@angular/animations": "~13.0.1",
+    "@angular/animations": "^14.3.0",
     "@angular/cdk": "^13.0.2",
-    "@angular/common": "~13.0.1",
-    "@angular/compiler": "~13.0.1",
-    "@angular/core": "~13.0.1",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
     "@angular/fire": "^7.2.0",
     "@angular/flex-layout": "^12.0.0-beta.34",
-    "@angular/forms": "~13.0.1",
+    "@angular/forms": "^14.3.0",
     "@angular/material": "13.0.2",
-    "@angular/platform-browser": "~13.0.1",
-    "@angular/platform-browser-dynamic": "~13.0.1",
-    "@angular/router": "~13.0.1",
-    "@angular/service-worker": "~13.0.1",
-    "@datorama/akita": "6.2.0",
-    "@datorama/akita-ng-router-store": "^6.0.0",
-    "@ngneat/hot-toast": "3.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
+    "@datorama/akita": "^8.0.1",
+    "@datorama/akita-ng-router-store": "^8.0.0",
+    "@ngneat/hot-toast": "^4.1.1",
     "@ngneat/overview": "^2.0.2",
     "@ngneat/until-destroy": "^8.0.4",
     "@ngx-pwa/local-storage": "13",
@@ -45,22 +45,22 @@
     "ng-for-repeat": "^1.0.1",
     "ngx-vibration": "^1.1.1",
     "rxfire": "^6.0.0",
-    "rxjs": "~6.6.0",
+    "rxjs": "^7.8.1",
     "tinycolor2": "^1.4.2",
     "tslib": "^2.1.0",
     "xml-js": "^1.6.11",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.1300.2",
-    "@angular-devkit/build-angular": "~13.0.2",
+    "@angular-devkit/architect": "0.1402.13",
+    "@angular-devkit/build-angular": "^14.2.13",
     "@angular-eslint/builder": "13.0.1",
     "@angular-eslint/eslint-plugin": "13.0.1",
     "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/schematics": "13.0.1",
+    "@angular-eslint/schematics": "^14.4.0",
     "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "~13.0.2",
-    "@angular/compiler-cli": "~13.0.1",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
     "@datorama/akita-ngdevtools": "^6.0.0",
     "@types/jasmine": "~3.6.0",
     "@types/lodash-es": "^4.17.4",
@@ -82,6 +82,6 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "open": "^7.0.3",
-    "typescript": "~4.4.4"
+    "typescript": "~4.8.4"
   }
 }
